DHFPROD-6362: Getting testDhsTests working
Added installDhfIntoDhs as a convenience for first running the DHF installer for DHS
