+++ marklogic-data-hub/src/test/java/com/marklogic/hub/dhs/installer/command/InstallIntoDhsCommandTest.java
@@ -149,54 +149,4 @@ public class InstallIntoDhsCommandTest extends AbstractHubCoreTest {
         assertEquals("basic", props.getProperty("mlJobAuth"));
         assertEquals("basic", props.getProperty("mlStagingAuth"));
     }
-
-    @Test
-    public void testUpdateDhsResourcePermissions() {
-        try {
-            runAsAdmin();
-            new UpdateDhsModulesPermissionsCommand(getHubConfig()).execute(newCommandContext());
-
-            GenericDocumentManager modMgr = getHubClient().getModulesClient().newDocumentManager();
-
-            DocumentMetadataHandle metadataHandle = new DocumentMetadataHandle();
-            modMgr.readMetadata("/marklogic.rest.resource/mlDbConfigs/assets/metadata.xml", metadataHandle);
-            DocumentMetadataHandle.DocumentPermissions perms = metadataHandle.getPermissions();
-
-            Assertions.assertEquals(DocumentMetadataHandle.Capability.UPDATE, perms.get("data-hub-environment-manager").iterator().next());
-            Assertions.assertEquals(DocumentMetadataHandle.Capability.READ, perms.get("rest-extension-user").iterator().next());
-
-            DocumentMetadataHandle moduleMetadataHandle = new DocumentMetadataHandle();
-            modMgr.readMetadata("/marklogic.rest.resource/mlBatches/assets/resource.sjs", moduleMetadataHandle);
-            DocumentMetadataHandle.DocumentPermissions modulePerms = moduleMetadataHandle.getPermissions();
-
-            Assertions.assertEquals(DocumentMetadataHandle.Capability.UPDATE, modulePerms.get("data-hub-environment-manager").iterator().next());
-            Assertions.assertNull(modulePerms.get("rest-admin-internal"));
-
-            moduleMetadataHandle = new DocumentMetadataHandle();
-            modMgr.readMetadata("/marklogic.rest.resource/mlBatches/assets/resource.xqy", moduleMetadataHandle);
-            modulePerms = moduleMetadataHandle.getPermissions();
-
-            Assertions.assertEquals(DocumentMetadataHandle.Capability.UPDATE, modulePerms.get("data-hub-environment-manager").iterator().next());
-            Assertions.assertNull(modulePerms.get("rest-admin-internal"));
-
-            moduleMetadataHandle = new DocumentMetadataHandle();
-            modMgr.readMetadata("/marklogic.rest.transform/mlGenerateFunctionMetadata/assets/transform.sjs", moduleMetadataHandle);
-            modulePerms = moduleMetadataHandle.getPermissions();
-
-            Assertions.assertEquals(DocumentMetadataHandle.Capability.UPDATE, modulePerms.get("data-hub-environment-manager").iterator().next());
-            Assertions.assertNull(modulePerms.get("rest-admin-internal"));
-
-            moduleMetadataHandle = new DocumentMetadataHandle();
-            modMgr.readMetadata("/marklogic.rest.transform/mlGenerateFunctionMetadata/assets/transform.xqy", moduleMetadataHandle);
-            modulePerms = moduleMetadataHandle.getPermissions();
-
-            Assertions.assertEquals(DocumentMetadataHandle.Capability.UPDATE, modulePerms.get("data-hub-environment-manager").iterator().next());
-            Assertions.assertNull(modulePerms.get("rest-admin-internal"));
-        } finally {
-            new DatabaseManager(getHubClient().getManageClient()).clearDatabase(HubConfig.DEFAULT_MODULES_DB_NAME);
-            installHubModules();
-            installHubArtifacts();
-        }
-
-    }
 }
