+++ marklogic-data-hub/src/test/java/com/marklogic/hub/flow/impl/RunFlowRunnerTestsWithoutProjectTest.java
@@ -31,7 +31,7 @@ public class RunFlowRunnerTestsWithoutProjectTest extends FlowRunnerTest {
 
         // Need to avoid instantiating this twice for the benefit of the MultipleJobs test in the parent class
         if (myFlowRunner == null) {
-            myFlowRunner = new FlowRunnerImpl(getHubConfig().getHost(), getHubConfig().getMlUsername(), getHubConfig().getMlPassword());
+            myFlowRunner = new FlowRunnerImpl(getHubClient());
         }
         flowRunner = myFlowRunner;
